vue-admin: refactor LoginView, RequestPasswordResetView and ResetPasswordView and removed the Form components
