+++ backend-neonella/vue-admin/src/views/ResetPasswordView.vue
@@ -1,28 +1,59 @@
 <script setup>
-import NeonellaLogo from '@/assets/images/neonella-logo.svg'
-import ResetPasswordForm from '@/components/ResetPasswordForm.vue'
+import { onMounted, ref } from 'vue'
+import GuestLayout from '@/components/GuestLayout.vue'
+import InputField from '@/components/ui/InputField.vue'
+
+const newPasswordRef = ref(null)
+
+onMounted(() => {
+  newPasswordRef.value.$el.focus()
+})
 </script>
 
 <template>
-  <div
-    class="flex min-h-screen w-screen items-center justify-center bg-[url('@/assets/images/login-bg.jpg')] bg-cover bg-center bg-no-repeat px-4 md:px-0"
-  >
-    <div
-      class="flex max-w-[400px] flex-1 -translate-y-20 flex-col items-center rounded-lg border bg-secondary p-10"
-    >
-      <img :src="NeonellaLogo" alt="Neonella Logo" class="h-32 w-32" />
-
-      <h1
-        class="text-center font-pd text-3xl/3 font-medium leading-none text-dark"
-      >
-        Reset Your Password
-      </h1>
-
-      <div class="mb-5 mt-5 w-full">
-        <ResetPasswordForm />
+  <GuestLayout title="Reset your password">
+    <form class="space-y-6" method="POST">
+      <div>
+        <label for="new-password" class="block text-sm/6 font-medium text-dark"
+          >New Password</label
+        >
+        <div class="mt-2">
+          <InputField
+            ref="newPasswordRef"
+            id="new-password"
+            name="password"
+            type="password"
+            required
+          />
+        </div>
       </div>
-    </div>
-  </div>
+
+      <div>
+        <label
+          for="confirm_password"
+          class="block text-sm/6 font-medium text-dark"
+          >Confirm Password</label
+        >
+        <div class="mt-2">
+          <InputField
+            id="confirm_password"
+            name="confirm_password"
+            type="password"
+            required
+          />
+        </div>
+      </div>
+
+      <div>
+        <button
+          type="submit"
+          class="flex w-full justify-center rounded-md bg-accent px-3 py-1.5 text-sm/6 font-semibold text-light shadow-sm hover:bg-accent/90 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-accent"
+        >
+          Reset Password
+        </button>
+      </div>
+    </form>
+  </GuestLayout>
 </template>
 
 <style lang="scss" scoped></style>
